scripts/content_engine/requirements.txt
supabase==2.3.4
langchain==0.1.0
langchain_openai==0.0.2
sympy==1.12
python-dotenv==1.0.0
